scenario_inspector/requirements.txt
PyQt5>=5.15.0
PyYAML>=6.0
python-dateutil>=2.8.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
pyserial>=3.5
opencv-python>=4.5.0